remove databaseType parameter for GlobalRuleBuilder
