+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/rule/builder/GlobalRuleBuilder.java
@@ -18,7 +18,6 @@
 package org.apache.shardingsphere.infra.rule.builder;
 
 import org.apache.shardingsphere.infra.config.RuleConfiguration;
-import org.apache.shardingsphere.infra.database.type.DatabaseType;
 import org.apache.shardingsphere.infra.metadata.ShardingSphereMetaData;
 import org.apache.shardingsphere.infra.metadata.user.ShardingSphereUser;
 import org.apache.shardingsphere.infra.rule.scope.GlobalRule;
@@ -39,10 +38,9 @@ public interface GlobalRuleBuilder<R extends GlobalRule, T extends RuleConfigura
      * Build global rule.
      *
      * @param mataDataMap mata data map
-     * @param databaseType database type
      * @param ruleConfig rule configuration
      * @param users users
      * @return global rule
      */
-    R build(Map<String, ShardingSphereMetaData> mataDataMap, DatabaseType databaseType, T ruleConfig, Collection<ShardingSphereUser> users);
+    R build(Map<String, ShardingSphereMetaData> mataDataMap, T ruleConfig, Collection<ShardingSphereUser> users);
 }
